skill points can be gathered and spent
